Struct darkredis::CommandList

pub struct CommandList<'a> { /* fields omitted */ }

A struct for defining commands manually, which allows for pipelining of several commands. If you need to only run one command, use Command, which has almost the same API.

Example

use darkredis::{CommandList, Connection};

let command = CommandList::new("LPUSH").arg(b"pipelined-list").arg(b"bar")
    .command("LTRIM").arg(b"pipelined-list").arg(b"0").arg(b"100");
let results = connection.run_commands(command).await.unwrap();

assert_eq!(results, vec![Value::Integer(1), Value::Ok]);

Methods

impl<'a> CommandList<'a>

pub fn new(cmd: &'a str) -> Self

Create a new command list starting with cmd.

pub fn arg<D>(self, data: &'a D) -> Self where
D: AsRef<[u8]>,

Consumes the command and appends an argument to it, builder style. Note that this will NOT create a new command for pipelining. That's what Commandlist::command is for.

See also

append_arg

pub fn args<D>(self, arguments: &'a [D]) -> Self where
D: AsRef<[u8]>,

Add multiple arguments from a slice, builder-style.

See also

append_args

pub fn command(self, cmd: &'a str) -> Self

Add a command to be executed in a pipeline, builder-style. Calls to Command::arg will add arguments from now on.

See also

append_command

pub fn append_args<D>(&mut self, arguments: &'a [D]) where
D: AsRef<[u8]>,

Append arguments from a slice.

See also

args

pub fn append_arg<D>(&mut self, data: &'a D) where
D: AsRef<[u8]>,

Mutate self by adding an additional argument.

See also

arg

pub fn append_command(&mut self, cmd: &'a str)

Append a new command to self.

See also

command

pub fn command_count(&self) -> usize

Count the number of commands currently in the pipeline