Struct darkredis::Connection

pub struct Connection { /* fields omitted */ }

A connection to Redis. Copying is cheap as the inner type is a simple, futures-aware, Arc<Mutex>, and will not create a new connection. Use a ConnectionPool if you want to use pooled conections. Every convenience function can work with any kind of data as long as it can be converted into bytes. Check the redis command reference for in-depth explanations of each command.

Methods

impl Connection

pub async fn connect<'_, A>(
address: A,
password: Option<&'_ str>
) -> Result<Self> where
A: ToSocketAddrs,

Connect to a Redis instance running at address. If you wish to name this connection, run the CLIENT SETNAME command.

pub async fn run_commands<'_, '_>(
&'_ mut self,
command: CommandList<'_>
) -> Result<Vec<Value>>

Run a series of commands on this connection.

pub async fn run_command<'_, '_>(
&'_ mut self,
command: Command<'_>
) -> Result<Value>

Run a single command on this connection.

pub async fn subscribe<'_, K>(
__arg0: Self,
channels: &'_ [K]
) -> Result<MessageStream> where
K: AsRef<[u8]>,

Consume self, and subscribe to channels, returning a stream of Messages. As of now, there's no way to get the connection back, nor change the subscribed topics.

pub async fn psubscribe<'_, K>(
__arg0: Self,
patterns: &'_ [K]
) -> Result<PMessageStream> where
K: AsRef<[u8]>,

Exactly like subscribe, but subscribe to patterns instead.

pub async fn publish<'_, C, M>(
&'_ mut self,
channel: C,
message: M
) -> Result<isize> where
C: AsRef<[u8]>,
M: AsRef<[u8]>,

Publish message to channel. Returns how many clients received the message.

pub async fn set<'_, K, D>(&'_ mut self, key: K, value: D) -> Result<()> where
K: AsRef<[u8]>,
D: AsRef<[u8]>,

Sets key to value.

pub async fn set_with_expiry<'_, K, D>(
&'_ mut self,
key: K,
data: D,
expiry: Duration
) -> Result<()> where
K: AsRef<[u8]>,
D: AsRef<[u8]>,

Deprecated since 0.5.0:
This function will be removed in 0.6.0. Please use set_and_expire_seconds or set_and_expire_ms instead.

Set key to value, and set key to expire in expiry's timeframe.

Deprecated

std::time::Duration consists of nanoseconds and seconds, which don't map well to either SET EX nor SET PX. This function only used the seconds part of the Duration, which may be confusing.

pub async fn set_and_expire_seconds<'_, K, D>(
&'_ mut self,
key: K,
data: D,
seconds: u32
) -> Result<()> where
K: AsRef<[u8]>,
D: AsRef<[u8]>,

Set the key key to data, and set it to expire after seconds seconds.

pub async fn set_and_expire_ms<'_, K, D>(
&'_ mut self,
key: K,
data: D,
milliseconds: u32
) -> Result<()> where
K: AsRef<[u8]>,
D: AsRef<[u8]>,

Set the key key to data, and set it to expire after milliseconds ms.

pub async fn expire_seconds<'_, K>(
&'_ mut self,
key: K,
seconds: u32
) -> Result<isize> where
K: AsRef<[u8]>,

Set key to expire seconds seconds from now.

pub async fn expire_ms<'_, K>(
&'_ mut self,
key: K,
seconds: u32
) -> Result<isize> where
K: AsRef<[u8]>,

Set key to expire milliseconds ms from now.

pub async fn expire_at_seconds<'_, K>(
&'_ mut self,
key: K,
timestamp: u64
) -> Result<isize> where
K: AsRef<[u8]>,

Set key to expire at unix timestamp timestamp, measured in seconds.

pub async fn expire_at_ms<'_, K>(
&'_ mut self,
key: K,
timestamp: u64
) -> Result<isize> where
K: AsRef<[u8]>,

Set key to expire at unix timestamp timestamp, measured in milliseconds.

pub async fn del<'_, K>(&'_ mut self, key: K) -> Result<isize> where
K: AsRef<[u8]>,

Delete key.

Return value

The number of deleted keys

pub async fn get<'_, K>(&'_ mut self, key: K) -> Result<Option<Vec<u8>>> where
K: AsRef<[u8]>,

Get the value of key.

pub async fn lpush<'_, K, V>(&'_ mut self, list: K, value: V) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Push a value to list from the left.

Return value

The number of elements in list

pub async fn lpush_slice<'_, '_, K, V>(
&'_ mut self,
key: K,
data: &'_ [V]
) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Like lpush, but push multiple values.

pub async fn rpush<'_, K, V>(&'_ mut self, list: K, value: V) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Push a value to list from the right.

Return value

The number of elements in list

pub async fn rpush_slice<'_, '_, K, V>(
&'_ mut self,
key: K,
values: &'_ [V]
) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Like rpush, but push multiple values through a slice.

pub async fn lpop<'_, K>(&'_ mut self, list: K) -> Result<Option<Vec<u8>>> where
K: AsRef<[u8]>,

Pop a value from a list from the left side.

Return value

The value popped from list

pub async fn rpop<'_, K>(&'_ mut self, list: K) -> Result<Option<Vec<u8>>> where
K: AsRef<[u8]>,

Pop a value from a list from the right side.

Return value

The value popped from list

pub async fn blpop<'_, '_, K>(
&'_ mut self,
lists: &'_ [K],
timeout: u32
) -> Result<Option<Vec<Vec<u8>>>> where
K: AsRef<[u8]>,

Pop a value from one of the lists from the left side. Block timeout seconds when there are no values to pop (timeout=0 means infinite)

Return value

Ok(Some((list,value))): name of the list and corresponding value

Ok(None): timeout (no values)

Err(err): there was an error

pub async fn brpop<'_, '_, K>(
&'_ mut self,
lists: &'_ [K],
timeout: u32
) -> Result<Option<Vec<Vec<u8>>>> where
K: AsRef<[u8]>,

Pop a value from one of the lists from the right side. Block timeout seconds when there are no values to pop (timeout=0 means infinite)

Return value

Ok(Some((list,value))): name of the list and corresponding value

Ok(None): timeout (no values)

Err(err): there was an error

pub async fn lrange<'_, K>(
&'_ mut self,
list: K,
from: isize,
to: isize
) -> Result<Vec<Vec<u8>>> where
K: AsRef<[u8]>,

Get a series of elements from list, from index from to to. If they are negative, take the index from the right side of the list.

pub async fn llen<'_, K>(&'_ mut self, list: K) -> Result<Option<isize>> where
K: AsRef<[u8]>,

Get the number of elements in list, or None if the list doesn't exist.

pub async fn lset<'_, K, V>(
&'_ mut self,
list: K,
index: usize,
value: V
) -> Result<()> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Set the value of the element at index in list to value.

pub async fn ltrim<'_, K>(
&'_ mut self,
list: K,
start: usize,
stop: usize
) -> Result<()> where
K: AsRef<[u8]>,

Trim list from start to stop.

pub async fn incr<'_, K>(&'_ mut self, key: K) -> Result<isize> where
K: AsRef<[u8]>,

Increment key by one.

Return value

The new value of key.

pub async fn incrby<'_, K>(&'_ mut self, key: K, val: isize) -> Result<isize> where
K: AsRef<[u8]>,

Increment key by val.

Return value

The new value of key

pub async fn incrbyfloat<'_, K>(&'_ mut self, key: K, val: f64) -> Result<f64> where
K: AsRef<[u8]>,

Increment key by a floating point value val.

Return value

The new value of key

pub async fn decr<'_, K>(&'_ mut self, key: K) -> Result<isize> where
K: AsRef<[u8]>,

Decrement key by a floating point value val.

Return value

The new value of key

pub async fn decrby<'_, K>(&'_ mut self, key: K, val: isize) -> Result<isize> where
K: AsRef<[u8]>,

Decrement key by val.

Return value

The new value of key

pub async fn append<'_, K, V>(&'_ mut self, key: K, val: V) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Append a string val to key.

Return value

The new size of key

pub async fn mget<'_, '_, K>(
&'_ mut self,
keys: &'_ [K]
) -> Result<Vec<Option<Vec<u8>>>> where
K: AsRef<[u8]>,

Get the string value for every key, or `None`` if it doesn't exist

pub async fn mset<'_, '_, '_, K, V>(
&'_ mut self,
keys: &'_ [K],
values: &'_ [V]
) -> Result<()> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Set multiple keys at once. If the number of keys and values are not equal, set all keys that have values and vice versa.

pub async fn exists<'_, K>(&'_ mut self, key: K) -> Result<bool> where
K: AsRef<[u8]>,

Returns true if a key has been previously set.