Struct darkredis::Connection

pub struct Connection { /* fields omitted */ }

A connection to Redis. Copying is cheap as the inner type is a simple, futures-aware, Arc<Mutex>, and will not create a new connection. Use a ConnectionPool if you want to use pooled conections. Alternatively, there's the deadpool-darkredis crate. Every convenience function can work with any kind of data as long as it can be converted into bytes. Check the Redis command reference for in-depth explanations of each command.

Implementations

impl Connection

pub async fn connect<A>(address: A) -> Result<Self> where
A: ToSocketAddrs,

Connect to a Redis instance running at address. If you wish to name this connection, run the CLIENT SETNAME command.

pub async fn connect_and_auth<A, P>(address: A, password: P) -> Result<Self> where
A: ToSocketAddrs,
P: AsRef<[u8]>,

Connect to a Redis instance running at address, and authenticate using password.

pub async fn run_command<'_, '_>(
&'_ mut self,
command: Command<'_>
) -> Result<Value>

Run a single command on this connection.

pub async fn run_command_with_buffer<'_, '_, '_>(
&'_ mut self,
command: Command<'_>,
buffer: &'_ mut Vec<u8>
) -> Result<Value>

Run a single command on this connection, using buffer for serialization. See run_commands_with_buffer for more details.

pub async fn run_commands<'_, '_>(
&'_ mut self,
command: CommandList<'_>
) -> Result<ResponseStream>

Run a series of commands on this connection, returning a stream of the results.

pub async fn run_commands_with_buffer<'_, '_, '_>(
&'_ mut self,
command: CommandList<'_>,
buf: &'_ mut Vec<u8>
) -> Result<ResponseStream>

Run a series of commands on this connection, using buffer for serialization. This prevents allocations as long as buffer is large enough from before. The buffer will be empty when this function returns.

pub async fn hdel<'_, K, F>(&'_ mut self, key: K, field: F) -> Result<bool> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,

Delete field from the hash set stored at key.

Return value

true when the field was deleted, false if it didn't exist

pub async fn hdel_slice<'_, '_, K, F>(
&'_ mut self,
key: K,
fields: &'_ [F]
) -> Result<isize> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,

Delete every field in fields from the hash set stored at key.

Return value

The number of deleted fields

pub async fn hexists<'_, K, F>(&'_ mut self, key: K, field: F) -> Result<bool> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,

Check if field exists in the hash set key.

pub async fn hget<'_, K, F>(
&'_ mut self,
key: K,
field: F
) -> Result<Option<Vec<u8>>> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,

Get the value of field in the hash set at key.

pub async fn hset<'_, K, F, V>(
&'_ mut self,
key: K,
field: F,
value: V
) -> Result<isize> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,
V: AsRef<[u8]>,

Set the value of field in the hash set stored at key to Value.

Return value

The number of added fields(will be 1 if field was created, 0 if it already existed).

pub async fn hsetnx<'_, K, F, V>(
&'_ mut self,
key: K,
field: F,
value: V
) -> Result<bool> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,
V: AsRef<[u8]>,

Set the value of field in the hash set stored at key to Value. If field already exists, do nothing.

Return value

true if field was set, false otherwise.

pub async fn hset_many<'_, '_, K>(
&'_ mut self,
key: K,
builder: MSetBuilder<'_>
) -> Result<isize> where
K: AsRef<[u8]>,

Set fields in the hash set at key to their values in builder. See Connection::mset for more information.

Return value

The number of added fields.

pub async fn hincrby<'_, K, F>(
&'_ mut self,
key: K,
field: F,
val: isize
) -> Result<isize> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,

Increment field in the hash set key by val.

Return value

The field value after the increment.

pub async fn hincrbyfloat<'_, K, F>(
&'_ mut self,
key: K,
field: F,
val: f64
) -> Result<f64> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,

Increment field in the hash set key by val, floating point version.

Return value

The field value after the increment.

pub async fn hkeys<'_, K>(&'_ mut self, key: K) -> Result<Vec<Vec<u8>>> where
K: AsRef<[u8]>,

Get the name of each hash field stored at key.

pub async fn hlen<'_, K>(&'_ mut self, key: K) -> Result<isize> where
K: AsRef<[u8]>,

Get the number of fields in the hash stored at key.

pub async fn hstrlen<'_, K, F>(&'_ mut self, key: K, field: F) -> Result<isize> where
K: AsRef<[u8]>,
F: AsRef<[u8]>,

Get the number of bytes in field in the hash set key

pub async fn hvals<'_, K>(&'_ mut self, key: K) -> Result<Vec<Value>> where
K: AsRef<[u8]>,

Get the value of each field in the hash field stored at key.

pub async fn ping<'_>(&'_ mut self) -> Result<()>

Send a PING to the server, returning Ok(()) on success.

pub async fn subscribe<'_, K>(
__arg0: Self,
channels: &'_ [K]
) -> Result<MessageStream> where
K: AsRef<[u8]>,

Consume self, and subscribe to channels, returning a stream of Message's. As of now, there's no way to get the connection back, nor change the subscribed topics.

pub async fn psubscribe<'_, K>(
__arg0: Self,
patterns: &'_ [K]
) -> Result<PMessageStream> where
K: AsRef<[u8]>,

Exactly like subscribe, but subscribe to channels matching patterns instead.

pub async fn publish<'_, C, M>(
&'_ mut self,
channel: C,
message: M
) -> Result<isize> where
C: AsRef<[u8]>,
M: AsRef<[u8]>,

Publish message to channel.

Return Value

Returns how many clients received the message.

pub async fn set<'_, K, D>(&'_ mut self, key: K, value: D) -> Result<()> where
K: AsRef<[u8]>,
D: AsRef<[u8]>,

Set key to value.

pub async fn set_and_expire_seconds<'_, K, D>(
&'_ mut self,
key: K,
data: D,
seconds: u32
) -> Result<()> where
K: AsRef<[u8]>,
D: AsRef<[u8]>,

Set the key key to data, and set it to expire after seconds seconds.

pub async fn set_and_expire_ms<'_, K, D>(
&'_ mut self,
key: K,
data: D,
milliseconds: u32
) -> Result<()> where
K: AsRef<[u8]>,
D: AsRef<[u8]>,

Set the key key to data, and set it to expire after milliseconds ms.

pub async fn expire_seconds<'_, K>(
&'_ mut self,
key: K,
seconds: u32
) -> Result<isize> where
K: AsRef<[u8]>,

Set key to expire seconds seconds from now.

pub async fn expire_ms<'_, K>(
&'_ mut self,
key: K,
seconds: u32
) -> Result<isize> where
K: AsRef<[u8]>,

Set key to expire milliseconds ms from now.

pub async fn expire_at_seconds<'_, K>(
&'_ mut self,
key: K,
timestamp: u64
) -> Result<isize> where
K: AsRef<[u8]>,

Set key to expire at Unix timestamp timestamp, measured in seconds.

pub async fn expire_at_ms<'_, K>(
&'_ mut self,
key: K,
timestamp: u64
) -> Result<isize> where
K: AsRef<[u8]>,

Set key to expire at Unix timestamp timestamp, measured in milliseconds.

pub async fn del<'_, K>(&'_ mut self, key: K) -> Result<isize> where
K: AsRef<[u8]>,

Delete key.

Return value

The number of deleted keys.

pub async fn del_slice<'_, '_, K>(&'_ mut self, keys: &'_ [K]) -> Result<isize> where
K: AsRef<[u8]>,

Delete every element in keys.

Return value

The number of deleted keys

pub async fn get<'_, K>(&'_ mut self, key: K) -> Result<Option<Vec<u8>>> where
K: AsRef<[u8]>,

Get the value of key.

pub async fn lpush<'_, K, V>(&'_ mut self, list: K, value: V) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Push a value to list from the left.

Return value

The number of elements in list

pub async fn lpush_slice<'_, '_, K, V>(
&'_ mut self,
key: K,
data: &'_ [V]
) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Like lpush, but push multiple values.

pub async fn rpush<'_, K, V>(&'_ mut self, list: K, value: V) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Push a value to list from the right.

Return value

The number of elements in list

pub async fn rpush_slice<'_, '_, K, V>(
&'_ mut self,
key: K,
values: &'_ [V]
) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Like rpush, but push multiple values through a slice.

pub async fn lpop<'_, K>(&'_ mut self, list: K) -> Result<Option<Vec<u8>>> where
K: AsRef<[u8]>,

Pop a value from a list from the left side.

Return value

The value popped from list

pub async fn rpop<'_, K>(&'_ mut self, list: K) -> Result<Option<Vec<u8>>> where
K: AsRef<[u8]>,

Pop a value from a list from the right side.

Return value

The value popped from list

pub async fn blpop<'_, '_, K>(
&'_ mut self,
lists: &'_ [K],
timeout: u32
) -> Result<Option<Vec<Vec<u8>>>> where
K: AsRef<[u8]>,

Pop a value from one of the lists from the left side. Block timeout seconds when there are no values to pop (timeout=0 means infinite)

Return value

Ok(Some((list,value))): name of the list and corresponding value

Ok(None): timeout (no values)

Err(err): there was an error

pub async fn brpop<'_, '_, K>(
&'_ mut self,
lists: &'_ [K],
timeout: u32
) -> Result<Option<Vec<Vec<u8>>>> where
K: AsRef<[u8]>,

Pop a value from one of the lists from the right side. Block timeout seconds when there are no values to pop (timeout=0 means infinite)

Return value

Ok(Some((list,value))): name of the list and corresponding value

Ok(None): timeout (no values)

Err(err): there was an error

pub async fn lrange<'_, K>(
&'_ mut self,
list: K,
from: isize,
to: isize
) -> Result<Vec<Vec<u8>>> where
K: AsRef<[u8]>,

Get a series of elements from list, from index from to to. If they are negative, take the index from the right side of the list.

pub async fn llen<'_, K>(&'_ mut self, list: K) -> Result<Option<isize>> where
K: AsRef<[u8]>,

Get the number of elements in list, or None if the list doesn't exist.

pub async fn lset<'_, K, V>(
&'_ mut self,
list: K,
index: usize,
value: V
) -> Result<()> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Set the value of the element at index in list to value.

pub async fn ltrim<'_, K>(
&'_ mut self,
list: K,
start: usize,
stop: usize
) -> Result<()> where
K: AsRef<[u8]>,

Trim list from start to stop.

pub async fn incr<'_, K>(&'_ mut self, key: K) -> Result<isize> where
K: AsRef<[u8]>,

Increment key by one.

Return value

The new value of key.

pub async fn incrby<'_, K>(&'_ mut self, key: K, val: isize) -> Result<isize> where
K: AsRef<[u8]>,

Increment key by val.

Return value

The new value of key

pub async fn incrbyfloat<'_, K>(&'_ mut self, key: K, val: f64) -> Result<f64> where
K: AsRef<[u8]>,

Increment key by a floating point value val.

Return value

The new value of key

pub async fn decr<'_, K>(&'_ mut self, key: K) -> Result<isize> where
K: AsRef<[u8]>,

Decrement key by a floating point value val.

Return value

The new value of key

pub async fn decrby<'_, K>(&'_ mut self, key: K, val: isize) -> Result<isize> where
K: AsRef<[u8]>,

Decrement key by val.

Return value

The new value of key

pub async fn append<'_, K, V>(&'_ mut self, key: K, val: V) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Append a string val to key.

Return value

The new size of key

pub async fn mget<'_, '_, K>(
&'_ mut self,
keys: &'_ [K]
) -> Result<Vec<Option<Vec<u8>>>> where
K: AsRef<[u8]>,

Get the string value for every key, or `None`` if it doesn't exist

pub async fn mset<'_, '_>(&'_ mut self, builder: MSetBuilder<'_>) -> Result<()>

Set every key in builder to their respective values.

Example

 use darkredis::{Connection, MSetBuilder};
 let mut connection = Connection::connect("127.0.0.1:6379").await.unwrap();
 let builder = MSetBuilder::new()
     .set(b"multi-key-1", b"foo")
     .set(b"multi-key-2", b"bar")
     .set(b"multi-key-3", b"baz");
 connection.mset(builder).await.unwrap();
 let keys = &[b"multi-key-1", b"multi-key-2", b"multi-key-3"];
 let results = connection.mget(keys).await.unwrap();
 assert_eq!(
     results,
     vec![
         Some(b"foo".to_vec()),
         Some(b"bar".to_vec()),
         Some(b"baz".to_vec())
     ]
 );

pub async fn exists<'_, K>(&'_ mut self, key: K) -> Result<bool> where
K: AsRef<[u8]>,

Returns true if a key has been previously set.

pub async fn sadd<'_, K, V>(&'_ mut self, key: K, value: V) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Adds new value to set specified by key.

pub async fn sadd_slice<'_, '_, K, V>(
&'_ mut self,
key: K,
values: &'_ [V]
) -> Result<isize> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Like sadd, but push multiple values.

pub async fn smembers<'_, K>(&'_ mut self, key: K) -> Result<Vec<Vec<u8>>> where
K: AsRef<[u8]>,

Return the members of a set specified by key.

pub async fn sismember<'_, K, V>(&'_ mut self, key: K, value: V) -> Result<bool> where
K: AsRef<[u8]>,
V: AsRef<[u8]>,

Returns true if value belongs to a set specified by key.